#79 refactor: Organize delimiter util function
