+++ src/components/SearchBar/SearchBar.tsx
@@ -5,7 +5,6 @@ import { Icon } from "@components/common/Icon";
 import useOutsideClick from "@components/hooks/useOutsideClick";
 import { useDebounce } from "@fineants/demolition";
 import { IconButton } from "@mui/material";
-import { removeWord } from "@utils/removeWord";
 import { ChangeEvent, useRef, useState } from "react";
 import styled from "styled-components";
 
@@ -154,7 +153,7 @@ function SearchItem({
     <StyledSearchItem onClick={onSearchItemClick}>
       <SearchItemName>
         <span>{value}</span>
-        {removeWord(searchResult.companyName, value)}
+        {(searchResult.companyName, value)}
       </SearchItemName>
       <label>{searchResult.tickerSymbol}</label>
     </StyledSearchItem>
