--- src/utils/removeWord.ts
@@ -1,4 +1,0 @@
-export function removeWord(text: string, wordToRemove: string) {
-  const regex = new RegExp(wordToRemove, "gi");
-  return text.replace(regex, "").trim();
-}
